src/main/java/ee/v22/model/solr/SearchResponse.java
package ee.v22.model.solr;

public class SearchResponse {

    private Response response;

    private ResponseHeader responseHeader;

    public Response getResponse() {
        return response;
    }

    public void setResponse(Response response) {
        this.response = response;
    }

    public ResponseHeader getResponseHeader() {
        return responseHeader;
    }

    public void setResponseHeader(ResponseHeader responseHeader) {
        this.responseHeader = responseHeader;
    }
}
